requirements.txt
livekit-agents>=1.2.0
livekit-plugins-groq>=0.1.0
livekit-plugins-deepgram>=0.1.0
livekit-plugins-noise-cancellation>=0.2.0
livekit-plugins-silero>=0.1.0
livekit-plugins-elevenlabs>=0.1.0
python-dotenv
livekit-agents[google]~=1.0